refactor: remove guide command (#1556)
