--- docs/guide/README.md
@@ -1,34 +1,0 @@
-# Contentful CLI - `guide` command
-
-A guide introducing basic concepts for working with Contentful.
-
-## What steps does the guide contain?
-
-### 1. Sign in to a new or existing account
-
-A shortcut to [contentful login](../login) to log the user in or offer a sign up
-form. This step will be skipped when already logged in.
-
-### 2. Create a Space to hold your content
-
-A shortcut to [contentful space create](../space/create) to create a new space.
-
-### 3. Create first content model and entries
-
-A shortcut to [contentful space seed](../space/seed) to set up a content model
-and add the first Content entries to the freshly created space.
-
-#### 4. Download app to display your content
-
-Download one of the [boilerplate example projects from Github](https://github.com/contentful?q=boilerplate&type=all&language=&sort=) 
-and add the required credentials to the project.
-
-#### 5. Run the app based on your environment
-
-No shortcut, the user is encouraged to run and hack the boilerplate he just downloaded.
-
-### Example
-
-```sh
-contentful guide
-```
