+++ lib/cmds/space_cmds/create.js
@@ -14,7 +14,7 @@ const LoggingSystem = require('../../core/event-handlers/logging')
 const createSpaceIntents = require('../../core/event-handlers/intents/create-space-handler')
 const createSpaceLogging = require('../../core/event-handlers/logging/create-space-handler')
 
-const { AbortedError } = require('../../guide/helpers')
+const { AbortedError } = require('../../utils/aborted-error')
 const { getHeadersFromOption } = require('../../utils/headers')
 
 module.exports.command = 'create'
