+++ lib/config.js
@@ -1,7 +1,6 @@
 module.exports = {
   noAuthNeeded: [
     'login',
-    'guide',
     'snake',
     'config add',
     'config list',
@@ -10,7 +9,6 @@ module.exports = {
   noSpaceIdNeeded: [
     'logout',
     'login',
-    'guide',
     'snake',
     'config add',
     'config list',
