+++ lib/utils/aborted-error.js
@@ -1,0 +1,2 @@
+class AbortedError extends Error {}
+module.exports.AbortedError = AbortedError
