+++ test/integration/cmds/snapshots/main.test.js.md
@@ -16,8 +16,6 @@ Generated by [AVA](https://ava.li).
        content-type  Manage and list your space content types␊
                                                                        [aliases: ct]␊
        extension     Manage and list your extensions␊
-       guide         A guide introducing basic concepts of working with␊
-                                  Contentful␊
        login         Login to Contentful␊
        logout        Logout from Contentful␊
        space         Manage and list your spaces␊
@@ -40,8 +38,6 @@ Generated by [AVA](https://ava.li).
        content-type  Manage and list your space content types␊
                                                                        [aliases: ct]␊
        extension     Manage and list your extensions␊
-       guide         A guide introducing basic concepts of working with␊
-                                  Contentful␊
        login         Login to Contentful␊
        logout        Logout from Contentful␊
        space         Manage and list your spaces␊
@@ -64,8 +60,6 @@ Generated by [AVA](https://ava.li).
        content-type  Manage and list your space content types␊
                                                                        [aliases: ct]␊
        extension     Manage and list your extensions␊
-       guide         A guide introducing basic concepts of working with␊
-                                  Contentful␊
        login         Login to Contentful␊
        logout        Logout from Contentful␊
        space         Manage and list your spaces␊
@@ -89,8 +83,6 @@ Generated by [AVA](https://ava.li).
        content-type  Manage and list your space content types␊
                                                                        [aliases: ct]␊
        extension     Manage and list your extensions␊
-       guide         A guide introducing basic concepts of working with␊
-                                  Contentful␊
        login         Login to Contentful␊
        logout        Logout from Contentful␊
        space         Manage and list your spaces␊
